fix: corrigi instalação de mensagem de instalação do Intellij

+++ ambiente_dev.sh
@@ -24,9 +24,14 @@ if [[ "Yy" == *"${confirmacao^^}"* ]] || [[ "" == *"${confirmacao^^}"* ]]; then
 
 fi
 
-read -p "Clique em Y caso não possua o docker na sua máquina para instala-lo, caso possua clique em N (Y/N): " confirmacao
+read -p "Clique em Y caso não possua o intellij na sua máquina para instala-lo, caso possua ou prefere usar sua própria IDE(configure por sua própria conta) clique em N (Y/N): " confirmacao
 
-sudo snap install intellij-idea-ultimate --classic
+
+if [[ "Yy" == *"${confirmacao^^}"* ]] || [[ "" == *"${confirmacao^^}"* ]]; then
+
+    sudo snap install intellij-idea-ultimate --classic
+
+fi
 
 
 read -p "Clique em Y caso não possua o docker na sua máquina para instala-lo, caso possua clique em N (Y/N): " confirmacao
